memperbaiki respons saat mengubah dan menghapus data supplier

+++ server/src/models/Supplier.js
@@ -13,17 +13,25 @@ module.exports = {
 		con.query(query, callback)
 	},
 	
-	update: (con, data, id, callback) => {		
-			const query = `UPDATE supplier SET
+	update: (con, data, id, res,callback) => {	
+			con.query(`SELECT * FROM supplier WHERE kd_supplier = ${id}`, (e, rows) => {
+				if(e) throw e
+				if(rows == 0) return res.send('id supplier tidak ditemukan.', 404)	
+				const query = `UPDATE supplier SET
 						nama_supplier = '${data.nama_supplier}',
 						alamat = '${data.alamat}'
 						WHERE kd_supplier = ${id}`
-		con.query(query, callback)
+				con.query(query, callback)
+			})
 	},
 
-	delete: (con, id, callback) => {
-		const query = `DELETE FROM supplier WHERE kd_supplier = ${id}`
-		con.query(query, callback)
+	delete: (con, id, res, callback) => {
+		con.query(`SELECT * FROM supplier WHERE kd_supplier = ${id}`, (e, rows) => {
+			if(e) throw e
+			if(rows == 0) return res.send('id supplier tidak ditemukan.', 404)
+			const query = `DELETE FROM supplier WHERE kd_supplier = ${id}`
+			con.query(query, callback)
+		})
 	}
 
 }
